knowledge-graph-service/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
python-arango>=7.5.0
psycopg2-binary>=2.9.7

# Redis
redis>=5.0.0

# Authentication
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
networkx>=3.2.0

# Text Processing
nltk>=3.8.1
spacy>=3.7.0

# HTTP Requests
httpx>=0.25.0
aiohttp>=3.9.0

# Async Processing
celery>=5.3.0

# Configuration
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Date/Time
python-dateutil>=2.8.2

# Utilities
pathlib2>=2.3.7
typing-extensions>=4.8.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
isort>=5.12.0
mypy>=1.6.0

# Optional: Machine Learning (if needed)
# scikit-learn>=1.3.0
# transformers>=4.35.0
# torch>=2.1.0